refactor advisory workspace replay evidence router
