+++ src/app/routers/advisory_workspace_version_lookups.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Path
+from fastapi import APIRouter
 
 from app.contracts.advisory_workspaces import AdvisoryWorkspaceEnvelopeResponse
 from app.middleware.correlation import correlation_id_var
@@ -22,29 +22,5 @@ async def list_saved_versions(
 ) -> AdvisoryWorkspaceEnvelopeResponse:
     return await advisory_workspace_service().list_saved_versions(
         workspace_id=workspace_id,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/{workspace_id}/saved-versions/{workspace_version_id}/replay-evidence",
-    response_model=AdvisoryWorkspaceEnvelopeResponse,
-    summary="Get Saved Advisory Workspace Replay Evidence",
-    description=(
-        "Returns replay evidence for a saved advisory workspace version from lotus-advise, "
-        "preserving source hashes and lifecycle continuity without Gateway-side inference."
-    ),
-)
-async def get_saved_version_replay_evidence(
-    workspace_id: str = WORKSPACE_ID_PATH,
-    workspace_version_id: str = Path(
-        ...,
-        description="Saved advisory workspace version identifier returned by lotus-advise.",
-        examples=["awv_001"],
-    ),
-) -> AdvisoryWorkspaceEnvelopeResponse:
-    return await advisory_workspace_service().get_saved_version_replay_evidence(
-        workspace_id=workspace_id,
-        workspace_version_id=workspace_version_id,
         correlation_id=correlation_id_var.get(),
     )
